docs: AI copilot implementation (#294)
Hey team,

We've been thoroughly enjoying Ragas. We stumbled upon you guys after
the OpenAI talk and we’ve been super impressed. Here at Mendable, we use
RAG pipelines all over so this has been super helpful!

Recently, when trying to implement, we stumbled into some questions. So,
we ended up ingesting the Ragas docs into Mendable.ai and started using
it to answer some of the questions that we had about the product.

We thought it did an awesome job answering most of them, so we figured
to PR and offer it as a feature for your docs. We implemented a small
button on the screen's bottom right which unveils an AI-powered chat
interface when clicked. There, users can ask questions about Ragas'
documentation and even debug code. The bot will learn with the
documentation daily so it will be always updated with recent doc
changes. (100% free btw)

We would love to hear your feedback! Super excited for what yall are
building and hopefully this can help a tiny bit :)

A bit of background: I am one of the co-founders of Mendable.ai. We are
supporters of open-source projects and we'd love to offer Mendable for
free, with the goal of helping yall out teach developers about your
product and answer users queries.

---------

Co-authored-by: rafaelmmiller <rafael@hypercube.be>