feat(metrics): skip embedding process when the corresponding weight is zero (#409)
This PR avoids calculating the embedding when assigning an embedding
weight of 0 to answer correctness

Currently, the correctness of the answer relies on both factuality and
semantic similarity. However, for most embedding models, the similarity
between any two embeddings will be larger than 0. This is equivalent to
introducing a bias into the rating system. Therefore, in practical use,
I tend to turn off similarity evaluation and only use factual
evaluation.

Do you think it would be better to allow embedding to be turned off like
this or to add an additional factuality metric?

---------

Co-authored-by: yuukidach <yuukidach@gmail.com>