Update Bring Your Own LLM Docs Link (#765)
This should correct the link to the page about using your own LLMs with
RAGAs. I haven't built/tested the docs locally to know if this fix
actually worked, I just referenced how the other hyperlinks in the docs
worked and went from there. Either way, this link is currently broken in
the public docs.

---------

Co-authored-by: Shahules786 <Shahules786@gmail.com>