Fix adapt issue for keyphrase extractor (#766)
Hi,
This fixes [issue
#625](https://github.com/explodinggradients/ragas/issues/625). This
issue is happening when after adaption testset generation is called.
This issue is caused by the fact, that after adaption the keyword
extractor returns nothing. This leads to failure later. However the
keywords are indeed extracted correctly (as can be seen when debugging)
but not returned properly from the function. The key "topics" doesn't
exist (after adaption) in the dict, instead it's "keyphrases". This PR
works for both with/without adaption.

---------

Co-authored-by: Sebastian Modlich <sebastian.modlich@telekom.de>
Co-authored-by: Shahules786 <Shahules786@gmail.com>