Add initial version of summarization metric (#990)
This PR adds an initial version of a metric for summarization use cases.

**Description**
For evaluating summaries generated by LLMs, we need a metric to consider
overall informativeness and conciseness of the summary. This metric has
been formulated to do the same.

**Formulation**
To determine if a summary is a good one or not, it is essential that it
contains all important information that is present in the original text.
We begin with generating `n` number of `'yes'/'no'` questions, for which
the original text strictly has the answer `yes`.
We then ask those questions to the generated summary, and compute the
`qa_score` as the fraction of correctly answered questions.

We then also take into account the length of the summary, because
otherwise this metric would be biased towards summaries that are same as
the original text, thus answering all questions correctly and obtaining
a score of 1. So we provide an option to add a penalty based on the
relative length of the summary.

For example, if the summary could answer `yes` to 3 out of 5
questions(`['yes', 'yes', 'no', 'yes', 'no']`), `qa_score` would be 0.6.
And if the summary length is 400 and original text length was 1000,
`relative_length` of summary would be 0.4.
Consider two summaries `S1` and `S2` for a text `T`, if `len(S1)=450`
and `len(S2)=600` and both have the same `qa_score`, we would prefer the
shorter summary, so I used `conciseness_score=1-(len(S)/len(T))` ,
shorter the summary, larger the `conciseness_score`.

Finally, I take the average of the two scores to return the
`summary_score`.

**Usage in RAGAS**(as per current code in the PR, subject to change)
```python
import datasets
import pandas as pd
from ragas.metrics import summarization_score
from ragas import evaluate

data = {}
data['text'] = ["Google LLC is an American multinational technology company that specializes in Internet-related services and products. Founded in 1998 by Larry Page and Sergey Brin, it is one of the Big Five technology companies. Google is known for its search engine, online advertising technologies, and various products like YouTube, Gmail, and Google Maps."]
data['summary'] = ["Google LLC, founded in 1998 by Larry Page and Sergey Brin, is an American multinational tech company known for its search engine and products like YouTube and Gmail."]


summary_eval_response = evaluate(
    datasets.Dataset.from_dict(data), 
    metrics=[summarization_score],)

```

**Discussions and Further thoughts**
- As far as I have checked, no other summarization metric codebase takes
relative length of the summary into account, so this could be useful to
promote shorter summaries.
- I am reading some papers and have some idea in mind to add more logic
to this metric, considering diversity(based on original text) in
summary, usefulness of summary etc. Will add to this PR or create new
ones.

@shahules786 @jjmachan Looking forward for review and discussions.

---------

Co-authored-by: Shahules786 <Shahules786@gmail.com>
Co-authored-by: jjmachan <jamesjithin97@gmail.com>