Fix: `adapt` function to translate cached `prompts` in place. (#889)
Address issue #888

It enables the following functionality in order to generate a testset in
a language other than English.

```python

from ragas.testset.prompts import testset_prompts
from ragas.testset.generator import TestsetGenerator

...
# some other imports 
...

# adapts all prompts to a given language, here german
for p in testset_prompts:
    p.adapt("german", translator_llm)

# run the test set generation
generator = TestsetGenerator(
        docstore=docstore,
        generator_llm=generator_llm,
        critic_llm=critic_llm,
        embeddings=embeddings
    )

dataset = generator.generate(
        test_size=TEST_SIZE,
        distributions=DISTRIBUTION
)
```

Without this fix, this would only work the first time. The first time,
when the prompts are translated the "adapt" method changes the prompt
instances (that is, in place). Therefore, the generate below uses the
translated prompts. However once the prompts are in the cache, running
adapt on all prompts doesn't change them "in place" and then `generate`
will access the original prompts completely in English. This PR makes
this approach to work also when the translated prompts are already
cached.

I also added the `testset_prompts` variable for convenience.

Co-authored-by: Ivan Herreros <ivan.herreros-alonso@inovex.de>