Fix test command in development.md to match Makefile configuration (#1195)
The development.md file previously suggested using make tests to run the
tests. However, the
[Makefile](https://github.com/explodinggradients/ragas/blob/main/Makefile#L28C1-L28C19)
does not include a target named tests, which results in the error make:
Nothing to be done for 'tests'. This commit updates the documentation to
reflect the correct command for running tests, ensuring alignment with
the existing Makefile configuration.