Add `show_progress` flag to Executor class
Hello there :wave: I noticed that there was no way to disable the progress bar and thought it'd nice to have the option to silence the `tqdm` call in the executor.

I've added the following changes:

Add `show_progress` flag to `Executor` class and update `evaluate` function to use it.

* **Executor Class Changes:**
  - Add `show_progress` flag to the `Executor` class in `src/ragas/executor.py`.
  - Use the `show_progress` flag to control the display of the progress bar.
  - If `show_progress` is set to False, disable `tqdm` usage for displaying progress bars.

* **Evaluate Function Changes:**
  - Add `show_progress` parameter to the `evaluate` function in `src/ragas/evaluation.py`.
  - Pass the `show_progress` flag to the `Executor` class during initialization in the `evaluate` function.