This commit implements the F-beta score metric (#1543)
for the AnswerCorrectness class. The beta parameter is introduced to
control the relative importance of recall and precision when calculating
the score. Specifically:

 - beta > 1 places more emphasis on recall.
 - beta < 1 favors precision.
- beta ==1 stands for the regular F1 score that can be interpreted as a
harmonic mean of the precision and recall.

Key Changes:
The method _compute_statement_presence is updated to calculate the
F-beta score based on true positives (TP), false positives (FP), and
false negatives (FN).

This ensures that we can balance between recall and precision, depending
on the task's requirements, by tuning the beta value.

source:
https://scikit-learn.org/1.5/modules/generated/sklearn.metrics.fbeta_score.html

---------

Co-authored-by: Shahules786 <Shahules786@gmail.com>