feat: replaced NLTK's implementation of BLEU with sacrebleu's implementation (#1744)
NLTK's implementation of BLEU is limited. In particular, only 20% of my
attempts to compute BLEU return a score because the number of candidate
and reference sentences are not the same, a requirement of NLTK's
implementation. The implementation of BLEU by
[sacrebleu](https://github.com/mjpost/sacrebleu) is recommended because
it is more robust. In my PR, I have modified _bleu_score.py to use
sacrebleu's implementation.