feat: refactor Ragas `embeddings` to maintain backward compatibility (#2184)
Lot of small choices made in this PR, I'll walk through why they're done
this way:

1. Added more tests!
1. **API Contracts**: I've tried to retain the existing API contract
which the library expects till we decide to duplicate it or change it in
any way.
1. **Retain extensibility**: Register things dynamically as much as you
can — so lot of Python tricks and less dictionaries of registries
approaches.
1. **Hide complexity**: I have hidden complexity from the end user in
several places. As an example, you can use both the modern and the
legacy API interface because we have an overloaded the
`embedding_factory` function which detects the expected response and
creates a corresponding object.
1. **Existing implementations**: I also tinkered with existing metrics
and other uses of Embedding to confirm that the new and old both work
together. It looks overloaded with multiple type-checks and what not
right now — because we're supporting two behaviors in the same
namespace! This will go away when we can deprecate the existing API
completely.