Update context_precision.md (#2262)
Correct misleading and wrong definition of context precision in docs,
with examples

## Issue Link / Problem Description
<!-- Link to related issue or describe the problem this PR solves -->
- Fixes #2013
- OR describe the issue: What problem does this solve? How can it be
replicated?

The current definition of context precision is misleading (see
discussion in linked issue). This PR is aimed at correcting the
definition.

## Changes Made
<!-- Describe what you changed and why -->
- Corrected the definition of context precision
- Added examples to illustrate the metric better
- 

## References
<!-- Link to related issues, discussions, forums, or external resources
-->
- Related issues:
https://github.com/explodinggradients/ragas/issues/2013
- Documentation: 
- External references: 
---
<!-- 
Thank you for contributing to Ragas! 
Please fill out the sections above as completely as possible.
The more information you provide, the faster your PR can be reviewed and
merged.
-->