total bin coverage for default_transform() in Knowledge Graph transformations (#1950)
# Problem

default_transform() uses token lengths up to 100k (0-100k interval) and
seperates it into three bins.
But for longer documents with token length >100k and 0 this function
raises the following:
```python
 raise ValueError(
            "Documents appears to be too short (ie 100 tokens or less). Please provide longer documents."
        )
```
Which covers the case of empty documents but also violates the
constraint >100k.

# Solution (Currently implemented)

I'm not sure with this solution but my first approach was to change the
last bin interval to `inf`. This solves the problem easily but could be
inefficient for very large documents.
```python
    bin_ranges = [(0, 100), (101, 500), (501, float("inf"))]
```

# Better Solution Proposal (Let's discuss this)

If the given document is larger than >100k tokens. Seperate the document
in half. And start the transformation again, until it fits into the
initial bin sizes.