Improved context precision documentation (#2266)
## Issue Link / Problem Description
<!-- Link to related issue or describe the problem this PR solves -->
- Fixes #[issue_number]
- OR describe the issue: What problem does this solve? How can it be
replicated

The existing documentation has a mistake, and in some cases, poor style
and grammar or is missing key information.

## Changes Made
<!-- Describe what you changed and why -->
- Provided separate example for explaining Context Precision as a metric
- Corrected the definition of the metric
LLMContextPrecisionWithoutReference
- Improved Grammar and style
- For NonLLMContextPrecisionWithReference, added packages to be
installed and also included an example of the distance algorithm used

---
<!-- 
Thank you for contributing to Ragas! 
Please fill out the sections above as completely as possible.
The more information you provide, the faster your PR can be reviewed and
merged.
-->